requirements.txt
numpy>=1.22.0
pyyaml>=6.0
requests>=2.28.0
sqlalchemy>=1.4.0
matplotlib>=3.5.0
pandas>=1.4.0
scikit-learn>=1.0.0
tqdm>=4.64.0
faiss-cpu>=1.7.0
pyarrow>=8.0.0
scipy>=1.9.0
seaborn>=0.11.0
